Working ProtoType
I know there are more things to work on in the code but I got the
action cards up and running in a general way and also saved the scores
to the players text file for as many games and he/she chooses to play.
I also added up the scores so that at the end of each game the players
scores will be totaled and added to the class/ file.