Debug
There were a few bugs, like not clearing the decks before a start to a
new game as well as a few more adding issues with the scoring, but I
fixed them in the scoring function and then at the end of main as well.
There might be a few more changes but nothing major.