Final Edits
Just a few final touches to edit some of the code where I noticed some
major bugs. Like putting the correct cards on the discard pile and a
few other minor things. I also added system(CLS) in some of the code
but it does not work in Xcode; however, i know it works in visual
studios.